requirements.txt
Pillow==11.1.0
pandas==2.2.3